python3 conversions

+++ FLUXES/mydtype.py
@@ -3,16 +3,16 @@
 import numpy as np
 
 
-trans_dt=np.dtype([('name','S12'),('type','S12'),('lon1',np.float),('lon2',np.float),\
-                                                ('lat1',np.float),('lat2',np.float),\
-                                                ('depth1',np.float),('depth2',np.float)])
+trans_dt=np.dtype([('name','S12'),('type','S12'),('lon1',np.float32),('lon2',np.float32),\
+                                                ('lat1',np.float32),('lat2',np.float32),\
+                                                ('depth1',np.float32),('depth2',np.float32)])
 
 
-ind2d   =np.dtype([('lat',np.int),('lon',np.int)])
+ind2d   =np.dtype([('lat',int),('lon',int)])
 
 
 # datatype for fluxes
 
-flux_dt =np.dtype([('adv-u',np.float),('adv-v',np.float),('adv-w',np.float),('sed-w',np.float),\
-                   ('hdf-x',np.float),('hdf-y',np.float),('zdf-z',np.float)])
+flux_dt =np.dtype([('adv-u',np.float32),('adv-v',np.float32),('adv-w',np.float32),('sed-w',np.float32),\
+                   ('hdf-x',np.float32),('hdf-y',np.float32),('zdf-z',np.float32)])
